make partial file2sqlite

+++ drop_db.js
@@ -1,5 +1,9 @@
 /**удаление файла базы sqlite**/
 var fs = require('fs');
-var filename = 'el.sqlite';
+var filename = process.argv[2];
+if ( filename == undefined ){
+	console.log('db_file not defined; Usage: drop_db <db_file>');
+	process.exit(0);
+}
 
 if ( fs.existsSync(filename) ) fs.unlinkSync(filename);
